requirements.txt
# This file is used by pip to install required python packages
# Usage: pip install -r requirements.txt

# Flask Framework
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Werkzeug==1.0.1

# Flask Packages
Flask-Login==0.4.0
Flask-Migrate==2.0.2
Flask-Script==2.0.5
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
Flask-User==1.0.1.5

# Automated tests
pytest==3.0.5
pytest-cov==2.4.0

#Neural Net
albumentations==0.4.6
appnope==0.1.0
backcall==0.1.0
certifi==2020.6.20
cycler==0.10.0
decorator==4.4.2
future==0.18.2
imageio==2.9.0
imbalanced-learn==0.7.0
imblearn==0.0
imgaug==0.4.0
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.16.0
joblib==0.16.0
jupyter-client==6.0.0
jupyter-core==4.6.3
kiwisolver==1.2.0
librosa==0.7.1
matplotlib==3.3.0
networkx==2.4
numba==0.48.0
numpy==1.19.1
opencv-python==4.3.0.36
opencv-python-headless==4.3.0.36
packaging==20.4
pandas==1.1.0
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prompt-toolkit==3.0.4
ptyprocess==0.6.0
Pygments==2.5.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.0
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.2
Shapely==1.7.0
six==1.14.0
sklearn==0.0
threadpoolctl==2.1.0
tifffile==2020.7.24
torch==1.6.0
torch-lr-finder==0.2.0
torchvision==0.7.0
tornado==6.0.4
tqdm==4.48.2
traitlets==4.3.3
wcwidth==0.1.8